Remove redundant toLong() conversion method call
Fixes a new warning introduced in Kotlin 2.3.

+++ uniffi_bindgen/src/bindings/kotlin/templates/RustBufferTemplate.kt
@@ -47,7 +47,7 @@ open class RustBuffer : Structure() {
 
     @Suppress("TooGenericExceptionThrown")
     fun asByteBuffer() =
-        this.data?.getByteBuffer(0, this.len.toLong())?.also {
+        this.data?.getByteBuffer(0, this.len)?.also {
             it.order(ByteOrder.BIG_ENDIAN)
         }
 }
